Finalizando/ComeçoAPI

+++ loja/resources/views/vendedores/index.blade.php
@@ -10,9 +10,9 @@
         <div class="pull-right">
 
 
-
+        @can('vendedores-create')
             <a class="btn btn-success" href="{{ route('vendedores.create') }}"> + Novo Vendedor</a>
-
+        @endcan
 
 
         </div>
@@ -48,17 +48,19 @@
     <td>
        <a class="btn btn-info" href="{{ route('vendedores.show',$vendedor->id) }}">Mostrar</a>
 
+       @can('vendedores-edit')
 
         <a class="btn btn-primary" href="{{ route('vendedores.edit',$vendedor->id) }}">Editar</a>
+        @endcan
 
-
+        @can('vendedores-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['vendedores.destroy', $vendedor->id],'style'=>'display:inline']) !!}
 
             {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
 
         {!! Form::close() !!}
 
-
+        @endcan
     </td>
   </tr>
